server/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
boto3==1.35.7
python-dotenv==1.0.1
pymongo==4.8.0
PyJWT==2.9.0
bcrypt==4.2.0
requests==2.32.3